+++ apps/api/src/app/messages/dtos/get-messages-requests.dto.ts
@@ -23,17 +23,14 @@ export class GetMessagesRequestDto {
   @IsOptional()
   transactionId?: string[];
 
-  // @ApiPropertyOptional({
-  //   type: [String],
-  //   description: 'Filter by exact context keys (format: "type:id")',
-  //   example: ['tenant:org-123', 'region:us-east-1'],
-  // })
-  // TODO: IS_CONTEXT_ENABLED
-  @ApiHideProperty()
+  @ApiPropertyOptional({
+    type: String,
+    isArray: true,
+    description: 'Filter by exact context keys (format: "type:id")',
+    example: ['tenant:org-123', 'region:us-east-1'],
+  })
   @IsOptional()
-  @IsArray()
-  @IsString({ each: true })
-  contextKeys?: string[];
+  contextKeys?: string[] | string;
 
   @ApiPropertyOptional({
     type: Number,