+++ apps/worker/src/app/workflow/usecases/update-job-status/update-job-status.usecase.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { InstrumentUsecase } from '@novu/application-generic';
-import { JobRepository } from '@novu/dal';
+import { JobEntity, JobRepository } from '@novu/dal';
 
 import { UpdateJobStatusCommand } from './update-job-status.command';
 
@@ -9,7 +9,7 @@ export class UpdateJobStatus {
   constructor(private jobRepository: JobRepository) {}
 
   @InstrumentUsecase()
-  public async execute(command: UpdateJobStatusCommand): Promise<void> {
-    await this.jobRepository.updateStatus(command.environmentId, command.jobId, command.status);
+  public async execute(command: UpdateJobStatusCommand): Promise<JobEntity | null> {
+    return this.jobRepository.updateStatus(command.environmentId, command.jobId, command.status);
   }
 }