+++ libs/internal-sdk/src/funcs/activityRequestsList.ts
@@ -110,7 +110,7 @@ async function $do(
     options: client._options,
     baseURL: options?.serverURL ?? client._baseURL ?? '',
     operationID: 'ActivityController_getLogs',
-    oAuth2Scopes: [],
+    oAuth2Scopes: null,
 
     resolvedSecurity: requestSecurity,