+++ libs/internal-sdk/src/funcs/layoutsCreate.ts
@@ -114,7 +114,7 @@ async function $do(
     options: client._options,
     baseURL: options?.serverURL ?? client._baseURL ?? '',
     operationID: 'LayoutsController_create',
-    oAuth2Scopes: [],
+    oAuth2Scopes: null,
 
     resolvedSecurity: requestSecurity,