+++ libs/internal-sdk/src/funcs/subscribersPreferencesUpdate.ts
@@ -128,7 +128,7 @@ async function $do(
     options: client._options,
     baseURL: options?.serverURL ?? client._baseURL ?? '',
     operationID: 'SubscribersController_updateSubscriberPreferences',
-    oAuth2Scopes: [],
+    oAuth2Scopes: null,
 
     resolvedSecurity: requestSecurity,