+++ libs/internal-sdk/src/funcs/topicsSubscriptionsList.ts
@@ -123,7 +123,7 @@ async function $do(
     options: client._options,
     baseURL: options?.serverURL ?? client._baseURL ?? '',
     operationID: 'TopicsController_listTopicSubscriptions',
-    oAuth2Scopes: [],
+    oAuth2Scopes: null,
 
     resolvedSecurity: requestSecurity,