+++ libs/internal-sdk/src/funcs/translationsUpload.ts
@@ -111,7 +111,7 @@ async function $do(
     options: client._options,
     baseURL: options?.serverURL ?? client._baseURL ?? '',
     operationID: 'TranslationController_uploadTranslationFiles',
-    oAuth2Scopes: [],
+    oAuth2Scopes: null,
 
     resolvedSecurity: requestSecurity,